Complete Access Control API Of BlockChain Service

+++ services/blockchain-service/src/main/java/com/omvaultchain/blockchain/controller/dto/VersionAtRequest.java
@@ -2,8 +2,10 @@ package com.omvaultchain.blockchain.controller.dto;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class VersionAtRequest {
     private String fileId;
-    private int versionNumber;
+    private BigInteger versionNumber;
 }
